backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
